ffserver: added function to synchronize the playback of multiple streams.
If the "Preroll" option for a file is non-zero, the start-position in the file will be relative to the start of the current hour. In this way it is possible to synchronize multiple file-streams to a single clock (if the streams are opened in the same hour). The only requirement is that all file streams start approximately at the same time. The files can be individually corrected using the "Preroll" configuration option, specified in seconds (floating point, can also be negative).