Fixed: set the "key_frame" flag of the AVFrame for the h264 codec (using MMAL).
The key_frame flag is not set when the AVFrame is decoded using the h264 MMAL accelerated codec. The flag is always 1 making it impossible to know if the decoded frame was generated from an I-frame packet. Unfortunately this information is not available from the MMAL decoded buffer. Therefore the PTS value of the last I-frame is stored and checked when the decoded frame arrives from the hardware decoder.