Added: option to specify time-delta for splitting in the HLS encoder.
This function is similar to the implementation in segment.c (introduced in commit aef9e94). When adding a delta of e.g. 100ms, the I-frame is allowed to arrive 100ms before the grid time for splitting the video files.

Example:
This makes splitting a video stream (GOP interval of 1 second) in 5 second pieces more reliable. Without this option, the I-frame to split will be missed if it is just a little -before- the grid time set by the segment length (even if it is 1 microsecond early). Once it is missed, the split function will wait for the next I-frame, that will arrive after 1 second. This creates two files, one of 4 seconds and one of 6 seconds, instead of 2 equal files of 5 seconds.