Fixed: only initialize the "start_pts" value if the incoming packet is a video packet.
The packet type should be checked before the start value is assigned. For example, if the first packet to arrive is an audio packet, the value of start-pts is in the timebase units of the audio stream. This will cause the stream to be split on every I-frame of the video stream since the pts value of the audio stream is typically much smaller than the pts value of the video stream.